+++ lib/src/models/models.dart
@@ -5,4 +5,5 @@ export 'grand_total_model.dart';
 export 'heartbeat.dart';
 export 'project.dart';
 export 'response_wrapper_model.dart';
+export 'summary.dart';
 export 'user_model.dart';